Speed up nvtaskset by skipping CUDA context creation if possible
The GPU needs to be on before the GPC-to-TPC mapping registers can
be read. The easiest way to power on the GPU is to create a CUDA
context, but this is fairly expensive.

In nvtaskset, some GPU-using task will likely already be running,
and so we can skip CUDA context creation in the common case.

Bug fixes:
- Delete the temporary CUDA context created by nvtaskset after
  it is done with it. Fixes bug where nvtaskset would leak this
  context into any program it launches.

Additional changes:
- Style fixes in libsmctrl.c
- Remove superfluous newlines from error() calls in nvtaskset.c